separating tag cache - wip
